Less magic numbers in inventories-Fix drafting station shift click

+++ src/main/java/net/shadowmage/ancientwarfare/automation/tile/torque/TileTorqueSingleCell.java
@@ -154,7 +154,7 @@ public abstract class TileTorqueSingleCell extends TileTorqueBase {
     @Override
     public NBTTagCompound getDescriptionTag() {
         NBTTagCompound tag = super.getDescriptionTag();
-        tag.setInteger("clientEnergy", (int) clientDestEnergyState);
+        tag.setInteger("clientEnergy", clientDestEnergyState);
         return tag;
     }
